feat: port api command to oclif, reduce create-api.ts to re-export shim
- Move all createApiCommand logic to src/commands/api.ts
- Add default-export Api class extending BaseCommand<void>
- --database flag uses options: ['mongodb','mongo','sqlite','sql'] for oclif validation
- Add examples block from Commander addHelpText
- Reduce src/commands/create-api.ts to export * from './api' shim
- create-api.test.js passes unmodified via shim import path

Refs #93
